Move WithIndicator to driver
